requirements.txt
quart
google-generativeai
python-dotenv
pytest
httpx 